requirements.txt
streamlit==1.29.0
pandas==2.1.4
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-cloud-firestore==2.14.0
google-cloud-aiplatform==1.38.1
sentence-transformers==2.2.2
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
openpyxl==3.1.2
plotly==5.18.0
